ansible-setup: avoid unnecesary apt cache updates
Some package installation (with latest) requires the apt cache to be updated.
If using the builtin mechanism for cache update, that would update the whole cache for all
packages and therefore (potentially) mark a bunch of packages for upgrade.

Prevent this by introducing a way to just upgrade the apt cache for a single repository.

Signed-off-by: Arturo Borrero Gonzalez <arturo.bg@arturo.bg>